test/web-platform-tests/to-upstream/css/cssom/cssimportrule-layer-false-positive.html

<!doctype html>
<title>@import layer detection should not match layer identifiers in supports() or media queries</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<link rel="help" href="https://drafts.csswg.org/css-cascade-5/#at-import">
<style id="supports-layer">
  @import "data:text/css," supports(layer);
</style>
<script>
"use strict";

test(() => {
  const rule = document.getElementById("supports-layer").sheet.cssRules[0];
  assert_equals(rule.layerName, null, "layerName should be null when layer only appears inside supports()");
}, "@import with supports(layer) should not set layerName");
</script>
